*** empty log message ***

+++ bundles/org.eclipse.swt/Eclipse SWT/win32/org/eclipse/swt/widgets/Display.java
@@ -2280,9 +2280,9 @@ public boolean post (Event event) {
 					case 127: inputs.wVk = (short) OS.VK_DELETE; break;
 					default : {
 						if (OS.IsWinCE) {
-							inputs.wVk = (short) OS.CharUpper ((short) key);
+							inputs.wVk = OS.CharUpper ((short) key);
 						} else {
-							inputs.wVk = (short) OS.VkKeyScan ((short) wcsToMbcs (key,0));
+							inputs.wVk = OS.VkKeyScan ((short) wcsToMbcs (key,0));
 							if (inputs.wVk == -1) return false;
 						}
 					}
